Rev2.01: Made conditional all parts of code specific to M0 Adalogger
Revision 2.00 of this code was written specifically for the Adafruit M0
Adalogger microcontroller. By popular demand, all lines of the code
specific to this chip are now wrapped in a conditional triggered by
USE_ADALOGGER define. Comment this define if you want to use another
MCU, but read the warning included in README before you do that!