Bug fix, motivated by issue #11
Although the n_last_peak_interval variable (the one that determines calculated heart rate) was properly initialized to the INIT_INTERVAL parameter, the first failure of the rf_signal_periodicity() routine was resetting it back to the FS value (a remnant of older code). For this reason, changing TYPICAL_HR to any other value than the 60 bpm default might have not resulted in desired effect.